videos_functions added
