+++ src/test.py
@@ -1,7 +1,11 @@
-import os
-from dotenv import load_dotenv
+import time
+import progressbar
 
-load_dotenv("secrets/.env")
 
-print(type(os.getenv("YT_PLAYLIST_ID")))
+video_lis = ["je", "suis", "un", "test"]
 
+
+with progressbar.ProgressBar(max_value=751.68) as bar:
+    for i, z  in enumerate(video_lis):
+        time.sleep(1)
+        bar.update(i+5.6875454564654564544)
