+++ src/videos.py
@@ -1,16 +1,13 @@
 import os
-
+import random
 import moviepy.editor as me
 from pathlib import Path
-from PIL import Image, ImageDraw, ImageFont, ImageFilter
-import textwrap
-BASE_DIR = Path().cwd()
+import time
+import progressbar
 
-# Parameter for thumbnail
-BACKGROUND_TINT_COLOR = (0, 0, 0)  # Black
-TRANSPARENCY = .40  # Degree of transparency, 0-100%
-TEXT_COLOR = (255, 255, 255)  # White
-FONT = 'Oswald.ttf'
+from dropbox_functions import _dropbox_download_file, _dropbox_list_files
+
+BASE_DIR = Path().cwd()
 
 
 def ajouter_audio_sur_video(path_videoclip, path_audioclip, loop=False):
@@ -39,101 +36,71 @@ def ajouter_audio_sur_video(path_videoclip, path_audioclip, loop=False):
     return f"Output/{nom_podcast}.mp4"
 
 
-def clip_transition_foudu(path_fodler, titre_video):
-    clip = []
-    for fichier_video in Path(path_fodler).iterdir():
-        clip.append(me.VideoFileClip(f"{fichier_video.parent}/{fichier_video.name}", audio=False, target_resolution=(720, 1280)).crossfadein(1))
 
-    clip_final = me.concatenate(clip, padding=-1, method="compose")
-    clip_final.write_videofile(f"Output/-{titre_video}.mp4")
-    return f"Output/-{titre_video}.mp4"
-
-
-def type_de_video_par_mot_cle(liste_titre_podcast):
-    dossier_mot_cle = {}
-
-    for mot in liste_titre_podcast:
-        if dossier_mot_cle.get(mot) is not None:
-            return dossier_mot_cle.get(mot)
-    return("Autre")
-
-
-def image_from_video(path_videoclip):
-    """ Télécharge l'image a 10 secondes de la video """
-
-    clip = me.VideoFileClip(path_videoclip)
-    duration = clip.duration
-    max_duration = int(duration) + 1
-    for i in range(max_duration):
-        if i == 10:
-            print(f"frame at {i} seconds")
-            frame = clip.get_frame(int(i))
-            os.makedirs("Thumbnails", exist_ok=True)
-            new_img_filepath = "Thumbnails/thumb.jpg"
-            new_img = Image.fromarray(frame)
-            new_img = new_img.resize((1280, 720))
-            new_img.save(new_img_filepath)
-            clip.close()
-            return new_img_filepath
-
-
-def draw_multiline_in_image(image_path, text, size_font=100, text_color=TEXT_COLOR, text_start_height=200):
-    OPACITY = int(255 * TRANSPARENCY)
-    print(len(text))
-    image = Image.open(image_path)
-    image = image.convert("RGBA")
-
-    # Font size according to the number of characters
-    if len(text) < 40:
-        size_font = 120
-    elif len(text) < 50:
-        size_font = 110
-    elif len(text) < 60:
-        size_font = 95
-    else:
-        size_font = 85
-
-    font = ImageFont.truetype(FONT, size_font)
-    image_width, image_height = image.size
-    y_text = text_start_height
-
-    width_text = 1
-    size_width_text = 0
-    while size_width_text < 1000:
-        width_text += 1
-        lines = textwrap.wrap(text.upper(), width=width_text)
-        size_width_text = font.getsize(lines[0])[0]
-
-    overlay = Image.new('RGBA', image.size, BACKGROUND_TINT_COLOR + (0,))
-    for line in lines:
-        line_width, line_height = font.getsize(line)
-
-        # Create piece of canvas to draw text on and blur
-        blurred = Image.new('RGBA', image.size)
-        draw = ImageDraw.Draw(blurred)
-        draw.text(((image_width - line_width) / 2, y_text),
-                  line, fill='black', font=font)
-        blurred = blurred.filter(ImageFilter.BoxBlur(7))
-
-        # Paste soft text onto background
-        image.paste(blurred,blurred)
-
-        draw_rect = ImageDraw.Draw(overlay)
-
-        draw_rect.text(((image_width - line_width) / 2, y_text),
-                  line, font=font, fill=text_color)
-        y_text += line_height
-    image = Image.alpha_composite(image, overlay)
-    image = image.convert("RGB")
-    image.save("result.jpg")
 
 
 ################################################
-# NEW FUNCTIONS
+# UPDATED FUNCTIONS
+
+def mix_video_list(dropbox_videos_path):
+    videos_list = _dropbox_list_files(dropbox_videos_path)['name']
+    random.shuffle(videos_list)
+    return videos_list
+
+
+def type_of_video_by_keyword(podcast_title):
+    string_title_to_list = podcast_title.split()
+    keyword_folders = {}
+    for keyword in string_title_to_list:
+        if keyword_folders.get(keyword) is not None:
+            return keyword_folders.get(keyword)
+    return("Autre")
+
+
+def download_video_same_length_as_audio(audio_path, dropbox_videos_path):
+    audio_duration = me.AudioFileClip(audio_path).duration
+    videos_list = mix_video_list(dropbox_videos_path)
+    video_duration = 0
+    print("Téléchargement des vidéos...")
+    with progressbar.ProgressBar(max_value=audio_duration+100) as bar:
+        for prefix, video in enumerate(videos_list, start=1):
+            video_path = f"tmp/Videos/{prefix}-{video}"
+            dropbox_path= f"{dropbox_videos_path}/{video}"
+            _dropbox_download_file(dropbox_path, video_path)
+            video_duration += me.VideoFileClip(video_path).duration
+            bar.update(video_duration)
+            if video_duration >= audio_duration:
+                return
+
+
+def video_editing(video_folder_path, video_title):
+    clip = []
+    for video_file in Path(video_folder_path).iterdir():
+        clip.append(me.VideoFileClip(f"{video_file.parent}/{video_file.name}", audio=False, target_resolution=(720, 1280)).crossfadein(1))
+
+    final_clip = me.concatenate(clip, padding=-1, method="compose")
+    final_clip.write_videofile(f"tmp/Output/{video_title}.mp4")
+    return f"tmp/Output/{video_title}.mp4"
+
 
 def create_video(podcast):
     #Download the podcast
+    os.makedirs("tmp/Podcast", exist_ok=True)
+    local_path = "tmp/Podcast/" + podcast['title'] + ".mp3"
+    _dropbox_download_file(podcast['path'], local_path)
+
     #Download the videos
+    os.makedirs("tmp/Videos", exist_ok=True)
+    dropbox_videos_path = f"/Video/{type_of_video_by_keyword(podcast['title'])}"
+    download_video_same_length_as_audio(local_path, dropbox_videos_path)
+
     #Video montage
+    os.makedirs("tmp/Output", exist_ok=True)
+    video_editing("tmp/Videos", podcast['title'])
+
     #Add audio in the video
-    pass
+    
+
+
+if __name__ == '__main__':
+    print(mix_video_list("/Video/Autre"))
